+++ src/queries/index.js
@@ -1,0 +1,3 @@
+
+export { default as useConfigurations } from './useConfigurations';
+export { default as useOkapiEnv } from './useOkapiEnv';